### 2022-02-09 #157 @labarba
@sbocce Hi! 👋
Thanks for your submission to JOSE. We'll be looking for a handling editor who will search for appropriate reviewers. But in the meantime, I want to say that you need to work a bit on the paper to flesh out the educational use of the software. Can you say more about how it is used in the classroom? We're looking for sufficient presentation so that another instructor may be able to reuse the software and/or educational content in their teaching.
