### 2022-02-09 #157 @sbocce
@labarba Hello, thank you for your message.
Sure, I will modify the paper following your suggestions and push the changes on the repository.
Just one thing: I see that the whedon code counting functionality has missed most of the code (all Fortran files). We previously had this problem and solved it switching from cloc to tokei.
