### 2022-02-10 f0324e2
Add educational use examples to the paper
