### 2022-02-10 #157 @sbocce
@labarba Hello, I have modified the paper introducing some examples to show how the software served as an educational tool, in the past. I'm glad to introduce further examples and/or clarifications if needed.
