### 2022-02-25 #157 @labarba
@ketch Hi! 👋
Would you consider volunteering to review for JOSE (The Journal of Open Source Education, sibling journal to JOSS). We are looking for reviewers with your skillset to check out this submission: "An Active Learning Laboratory Manual for Teaching a Contemporary Undergraduate Operating System Course" [View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00157/jose.00157/10.21105.jose.00157.pdf)
Have a look and let me know! Thank you
