### 2022-02-26 #163 @whedon
[REVIEW]: Hyper2D: a finite-volume solver for hyperbolic PDEs and hypersonic flows
**Submitting author:** @sbocce (<a href="http://orcid.org/0000-0003-2276-9606">Stefano Boccelli</a>)
**Repository:** <a href="https://github.com/sbocce/hyper2D" target ="_blank">https://github.com/sbocce/hyper2D</a>
**Version:** v1.0.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
