### 2022-02-28 sbocce/hyper2D#1 @ketch
Provide installation instructions
The README file does explain part of the installation process, but other parts are missing and the instructions are not clearly labeled.  I think just a section header ("Installation") and a few lines (clone from github, change directory, make) would suffice.
