### 2022-02-28 sbocce/hyper2D#2 @ketch
Document dependencies
Please indicate clearly what is required before installing / running this software (e.g., Fortran compiler, CUDA compiler, other libraries).
