### 2022-02-28 sbocce/hyper2D#3 @ketch
Add contribution guidelines
To save time, it would be very helpful if you read through what is expected for publication in JOSE and complete all the missing items, and then ping me to continue with the review.  You can see what is needed for example [here](https://github.com/openjournals/jose-reviews/issues/163).
