### 2022-03-01 sbocce/hyper2D#3 @sbocce
Hello @ketch, I have modified the paper & repository following the instructions. Here's a list of modifications and replies to the instructions.

Functionality:
- Installation -> I've introduced installation instructions in the main README and in the github wiki
- Functionality -> up to the reviewers I guess
- Performance -> there are no claims on performance in the manuscript, so I guess this is an automatic check

Documentation:
- A statement of need -> discussed in the intro of the wiki and the README
- Installation instructions -> I've introduced a "Software requirements" section in the main README and in the github wiki
- Example usage -> there is a sample simulation in the MS (to show the functionality), please let me know if this is not enough.
- Functionality documentation -> all I have prepared on this points are wiki pages to discuss the different modules of the solver (global variables, grid, PDE, numerical solvers)
- Tests -> I have prepared **no automated tests**. Actually, I had previously submitted Hyper2D to JOSS and the reviewers kindly suggested me to submit to JOSE instead for this very reason, that there are no tests and the software is mostly educational.
- Community guidelines -> I have introduced a "Contributing and reporting bugs" section in the github wiki

Software paper
- Authors -> done
- A statement of need -> done
- References -> done

Let me know if you need some further action on my side.
