### 2022-04-17 sbocce/hyper2D#3 @ketch
It seems that the main documentation is the Github wiki, but I don't see how anyone would possibly know to look there, except by reading this issue thread.  Perhaps you should state in some prominent places that the wiki is the documentation, with a link.
