### 2022-04-17 sbocce/hyper2D#3 @ketch
I find it very strange that the package targets beginners, but doesn't give any suggestions on how to plot the output.  Your introductory text mentions "good looking fluid dynamics simulations", but your package does not produce anything "good looking", right?

Additionally, saying that the output is "a simple dat file" is not helpful, since this doesn't mean anything (at least, to me).  Is it ASCII?  Is it binary?  What is the format?  How should one read it?
