### 2022-04-17 #163 @ketch
I have to say respectfully that I don't understand how this package is intended to be an educational resource.  How does one use it to learn?

It consists of examples of implementations of various algorithms for various problems.  But those problems and algorithms are not documented within the package.  In fact, there aren't even citations to other resources where the user could find out what equations are being solved and what discretizations are being used.  Perhaps I am missing something?
