### 2022-04-17 sbocce/hyper2D#3 @ketch
In the installation instructions, asking the user as a first step to browse the directory structure and find one with a Makefile is okay, but not great in my opinion.  It would be much better if you listed the directory for each package in the README.
