### 2022-04-17 #163 @labarba
From the submission cover message, I understand that this package was originally submitted to JOSS. I suspect that the editors there assessed that its research application is limited and suggested that probably JOSE was a better venue. The author's cover message says about the package that

> It targets (1) students learning numerical methods or needing a software for running computations for their MSc thesis, and (2) researchers in need of a simple code that is yet general enough for implementing complex PDEs.

If reuse of the package in teaching contexts needs additional documentation, that is something the authors should add in revision.
