### 2022-04-17 #163 @ketch
Well, after reading the JOSE guidelines and looking at the checklist here, I'm completely confused.  As far as I can tell, there are no "lessons" or pedagogical explanations of any kind in this package, but the reviewer checklist also doesn't mention anything of that kind.  What I can say is that I teach a course where this package would be relevant, but I don't see how it would be helpful to a student, given the lack of any pedagogical material.

The reviewer guidelines do ask me to look for tests, while the author claims that JOSS sent him here so he could get his paper accepted without needing to write tests.  I think there are some inconsistencies and ambiguities regarding the journal, and I don't feel capable of refereeing papers for it in light of that.  I have provided as much feedback as I could (here and in issues in the repository).  I hope that is helpful, and that you'll understand that I respectfully decline to play a further role in the evaluation of this submission.
