### 2022-04-17 #163 @labarba
JOSE admits two types of submissions: learning modules, and software submissions. Depending on the article type chosen by the author on submission, the Review issue will display one of two checklists. In this case, the checklist is for a software submission.

Our documentation does mention the [Specific requirements for software submissions](https://openjournals.readthedocs.io/en/jose/review_criteria.html#specific-requirements-for-software-submissions), including testing.

The [paper](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain) should "describe the functionality of the software, usage and recent experience of use in teaching and learning situations." The journal does seek accepted submissions to be geared for _reuse_ of the software (or lessons, in the case of learning modules), so the author should certainly address this.
