### 2022-04-17 #163 @labarba
>  the author claims that JOSS sent him here so he could get his paper accepted without needing to write tests

Where did you see that? Certainly that would be the wrong reason to transfer a submission from JOSS to JOSE. The right reason is that the software is not for research but for pedagogical uses.
