### 2022-04-17 #163 @ketch
https://github.com/sbocce/hyper2D

> Tests -> I have prepared no automated tests. Actually, I had previously submitted Hyper2D to JOSS and the reviewers kindly suggested me to submit to JOSE instead for this very reason, that there are no tests and the software is mostly educational.
