### 2022-04-17 #163 @mandli
I am in general agreement with @ketch.  At the very least the educational aspect/use of the software is lacking.
