### 2022-04-17 #163 @labarba
@sbocce — I'm seeing from the reviewers' comments (and exploring your repo) that this submission needs quite some work before it can be considered for publication. For the software to be reusable in teaching and learning contexts, it is essential that you provide documentation beyond raw-text README files in each folder. As it is, you seem to be expecting of users to "dig through" the repository to find out what they could do with these codes.

Most JOSE submissions are of the learning module category, but here is one in the software category:

> Ogbureke, K. U., (2021). mLEARn: An Implementation of Multi-layer Perceptron in C++. _Journal of Open Source Education_, 4(41), 59, https://doi.org/10.21105/jose.00059

The paper describes some use cases in the context of teaching and learning, and the software repository offers a well-formatted README with user instructions, examples, and an API doc in PDF. Users need these entry points to adopt a new open-source software package.

At this juncture, we have two options: we can add a `paused` label and you can come back and ping us when you have done more work in response to our input here, or we can withdraw the submission and you can resubmit when the software and documentation is in better shape to target reuse.
