### 2022-04-17 #163 @sbocce
Hello @labarba, hello @ketch & @mandli, thank you for the review & observations.

> Where did you see that? Certainly that would be the wrong reason to transfer a submission from JOSS to JOSE. The right reason is that the software is not for research but for pedagogical uses.

Just to clarify: once I submitted to JOSS some months ago, the reviewers pointed to me that rather than being a "full software package aimed at production use" aka JOSS target (and thus being shipped with automated google tests and all fancy things), it is instead a software more aimed towards educational use. So, they pointed me to JOSE, that I didn't know back then, and we thought that your venue is closer to our aims. So we came here.

> At this juncture, we have two options: we can add a paused label and you can come back and ping us when you have done more work in response to our input here, or we can withdraw the submission and you can resubmit when the software and documentation is in better shape to target reuse.

I would be happy to implement the said changes to the repository. I see that @ketch requested to get out of the review loop though. Anyway, if you want to add a paused label, I'll ping you once I have the submission ready for further review.
