### 2022-04-19 #163 @labarba
I'm adding the link here to the conversation over at JOSS, for future reference:

https://github.com/openjournals/joss-reviews/issues/4049#issuecomment-1013249736
