### 2022-04-19 #163 @labarba
☝️ note that nowhere did the JOSS editors imply that lack of testing was OK with us
