### 2022-04-19 #163 @sbocce
Hi @labarba thanks for your help.

Quick question about the type of tests admitted by JOSE:
> Tests: Are there automated tests **or** manual steps described so that the function of the software can be verified?

I do have the "manual steps described" for running the code. Do you think that adding a folder with the expected results in form of dat files and/or VTK and/or images will do?
