### 2022-04-19 #163 @labarba
@sbocce — My assessment is that your software submission is not ready for publication, for a number of reasons. If the reviewer(s) when looking through the submission realize that it would take an inordinate amount of work on their part to shepherd the necessary revisions to bring it up to publishable standards, it is not surprising that they would renege on their commitment to review. I believe this may be the case here.

It looks to me like you wrote this code as part of your PhD work and you are now looking to get a publication from it. This is what JOSS and JOSE were founded to provide: a publication venue for software written as part of a scholarly endeavor. But to be eligible for publication, the software needs to adhere to minimum standards that enable reusability: the code should be immediately useful to others.

For JOSE, the software should be useful in teaching and learning settings. If you did not _write_ the software with these uses in mind (you wrote it for your own purposes during your PhD), then you would have to invest the work to _make it usable_ for education.

In open-source software, users expect documentation, ways to check for correctness (tests), and good design (modularity, separation of concerns, low code redundancy, etc.). Our submission guidelines are not exhaustive, because they cannot be, and we do rely on community standards. The JOSS editors and the reviewers here saw that your software is not up to standards, and that the effort required on _their part_ to help you get there are substantial.

Your question [above](https://github.com/openjournals/jose-reviews/issues/163#issuecomment-1102170198) also leads me to the assessment that it could take a lot of effort on _our part_ if you need us to tell you how to provide checks for code correctness. (Certainly, images of expected results are not enough. Data files are not enough: there should be at least scripts that compare the golden files to the user's output.)

If you are still interested in pursuing a publication, I suggest you first work on the documentation (a wiki is not what users expect), then look at your code design for reusability (modify and edit docs as needed), and develop a couple of examples with testable expected output (perhaps read about code testing: you may not want to include unit tests, but you at least want regression tests).
