### 2022-04-19 #163 @sbocce
@labarba hi, thanks for your detailed reply.

Right, I see your point. Indeed, it seems pointless to keep working on the present submission, given the situation. If you could please just go ahead and withdraw my application, it would be amazing.

Once again thank you to both yourself and @ketch and @mandli for taking the time to check out my submission.
